more logging, minor bugfixes, getting closer
darcs-hash:20050423160110-57254-57542582c88eaccdca662e10d264008a54b9b110.gz

+++ src/udpsocket.cpp
@@ -124,7 +124,7 @@ void UDPSocket::send(NetPacket *packet)
 
 int UDPSocket::is_reply_to(struct sockaddr_in *addr)
 {
-    return (((short)their_addr.sin_port == (short)addr->sin_port)
+    return ((their_addr.sin_port == addr->sin_port)
             &&
             (their_addr.sin_addr.s_addr == addr->sin_addr.s_addr) );
 }
